+++ google-books-application/src/app/modules/books/books-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { BookDetailComponent } from './components/book-detail/components/detail-page/book-detail.component';
 import { AuthGuard } from '../../core/auth/guards/auth.guard';
 import { BooksFavoriteComponent } from './components/books-favorite/books-favorite.component';
 import { BooksSearchComponent } from './components/books-search/components/overview-page/books-search.component';
@@ -14,11 +13,6 @@ const routes: Routes = [
   {
     path: 'favorite',
     component: BooksFavoriteComponent,
-    canActivate: [AuthGuard],
-  },
-  {
-    path: 'book-detail/:id',
-    component: BookDetailComponent,
     canActivate: [AuthGuard],
   },
 ];