+++ google-books-application/src/app/modules/books/components/books-favorite/books-favorite.component.ts
@@ -1,7 +1,44 @@
-import { Component } from '@angular/core';
+import { Component, DestroyRef, OnInit, inject } from '@angular/core';
+import { BookCollection } from '../../models/book-collection.interface';
+import { BooksService } from '../../services/books.service';
+import { LoaderService } from '../../../../shared/services/loader.service';
+import { finalize } from 'rxjs';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 @Component({
   selector: 'app-books-favorite',
   templateUrl: './books-favorite.component.html',
   styleUrls: ['./books-favorite.component.scss'],
 })
-export class BooksFavoriteComponent {}
+export class BooksFavoriteComponent implements OnInit {
+  protected bookCollection: BookCollection | null = null;
+  private readonly destroyRef: DestroyRef = inject(DestroyRef);
+
+  constructor(
+    private booksService: BooksService,
+    private loaderService: LoaderService
+  ) {}
+
+  ngOnInit(): void {
+    this.fetchFavoriteBooks();
+  }
+
+  protected fetchFavoriteBooks(): void {
+    this.loaderService.start();
+    this.booksService
+      .getFavoriteBooks()
+      .pipe(
+        finalize(() => {
+          this.loaderService.stop();
+        }),
+        takeUntilDestroyed(this.destroyRef)
+      )
+      .subscribe({
+        next: (response) => {
+          if (!response) {
+            return;
+          }
+          this.bookCollection = response;
+        },
+      });
+  }
+}