+++ google-books-application/src/app/modules/books/constants/favorite-operation-message.const.ts
@@ -1,0 +1,4 @@
+export const FAVORITE_OPERATION_MESSAGE = {
+  ADDED: 'The book was successfully added to "favorite"',
+  REMOVED: 'The book was successfully removed from "favorite"',
+};