+++ google-books-application/src/app/shared/components/book-list-items/book-list-items.component.ts
@@ -1,9 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-book-list-items',
-  templateUrl: './book-list-items.component.html',
-})
-export class BookListItemsComponent {
-  @Input({ required: true }) items?: string[];
-}