+++ google-books-application/src/app/shared/shared.module.ts
@@ -2,8 +2,6 @@ import { CommonModule, NgOptimizedImage } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { BookCoverComponent } from './components/book-cover/book-cover.component';
-import { BookListItemsComponent } from './components/book-list-items/book-list-items.component';
 import { LoaderComponent } from './components/loader/loader.component';
 import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
 import { TruncateTextPipe } from './pipes/truncate-text.pipe';
@@ -12,13 +10,7 @@ import { LoaderService } from './services/loader.service';
 import { MaterialModule } from '../modules/material/material.module';
 
 @NgModule({
-  declarations: [
-    LoaderComponent,
-    PageNotFoundComponent,
-    BookListItemsComponent,
-    TruncateTextPipe,
-    BookCoverComponent,
-  ],
+  declarations: [LoaderComponent, PageNotFoundComponent, TruncateTextPipe],
   imports: [
     CommonModule,
     ReactiveFormsModule,
@@ -29,14 +21,12 @@ import { MaterialModule } from '../modules/material/material.module';
   ],
   exports: [
     LoaderComponent,
-    BookListItemsComponent,
     PageNotFoundComponent,
     TruncateTextPipe,
     CommonModule,
     ReactiveFormsModule,
     FormsModule,
     RouterModule,
-    BookCoverComponent,
     MaterialModule,
     NgOptimizedImage,
   ],